feature/back/AMS-2

+++ ams_backend/src/main/java/com/gate_software/ams_backend/repository/CheckOutRepository.java
@@ -3,5 +3,5 @@ package com.gate_software.ams_backend.repository;
 import com.gate_software.ams_backend.entity.CheckOutRecords;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CheckOutRepository extends JpaRepository<CheckOutRecords, Long> {
+public interface CheckOutRepository extends JpaRepository<CheckOutRecords, Integer> {
 }
