+++ ams_backend/src/main/java/com/gate_software/ams_backend/repository/DayRepository.java
@@ -3,5 +3,5 @@ package com.gate_software.ams_backend.repository;
 import com.gate_software.ams_backend.entity.Day;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DayRepository extends JpaRepository<Day, Long> {
+public interface DayRepository extends JpaRepository<Day, Integer> {
 }
